1441 Introduce transformators parameter in modules (#826)
https://github.com/insightsengineering/teal/issues/1441

Included `transformators` parameter in modules. This is passed to
`teal::module` that does assertions and checks on this parameter.

Did not introduce for `tm_data_table`, `tm_file_viewer` and
`tm_front_page` as we also didn't introduce `decorators` for those
modules. We can always introduce, if there is a user request, but I
don't think there will ever be.

Lastly, a food for thought: default `decorators` value is `NULL` where
default `transformators` value is `list()`. Maybe we can introduce a
check, that allows `transformators` to be `NULL` on default, and which
changes `NULL` to `list()`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Lluís Revilla <185338939+llrs-roche@users.noreply.github.com>