Fix tm_variable_browser: report plot (#941)
# Pull Request

Fixes #936

Generates code inside the qenv for it to be reproducible and make the
module functional.

I tried to make minimal changes but at the same time to make it fast to
render by adding new `req()` to invalidate reactives.

Note that the code created for the plot has some changes as it has the
construct for the new S7 labels. Maybe there are other issues with it.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>