Replace table by plot on tm_missing_data (#922)
# Pull Request

<!--- Replace `#nnn` with your issue link for reference. -->

Fixes #495 

Replaces the table by a plot on "By Variables Levels" tab. 
The colors match those in the other plots of the module.
It also required to add a "new" dependency, formatters, which is a
dependency of tern, rlistings, rtable. I added it on Suggests.

There is an extra commit to update the lintr names

```r
data <- within(teal_data(), {
  ADSL <- rADSL
  ADRS <- rADRS
})
join_keys(data) <- default_cdisc_join_keys[names(data)]

init(
  data = data,
  modules = modules(tm_missing_data())
) |> runApp()
```

---------

Signed-off-by: Lluís Revilla <185338939+llrs-roche@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: André Veríssimo <211358+averissimo@users.noreply.github.com>